[core] Merge light_properties.hpp into light_impl.hpp
