+++ src/mbgl/style/light.cpp
@@ -3,7 +3,6 @@
 #include <mbgl/style/light.hpp>
 #include <mbgl/style/light_impl.hpp>
 #include <mbgl/style/light_observer.hpp>
-#include <mbgl/style/light_properties.hpp>
 
 namespace mbgl {
 namespace style {
