+++ src/mbgl/style/light.cpp.ejs
@@ -6,7 +6,6 @@
 #include <mbgl/style/light.hpp>
 #include <mbgl/style/light_impl.hpp>
 #include <mbgl/style/light_observer.hpp>
-#include <mbgl/style/light_properties.hpp>
 
 namespace mbgl {
 namespace style {
